Move Navbar component to dedicated folder
Relocated the Navbar component and its CSS to components/navbar/. Updated all imports to reference the new path. This improves project structure and organization.
