+++ app/(additional)/layout.tsx
@@ -1,4 +1,4 @@
-import { Navbar } from "@/components/navbar";
+import { Navbar } from "@/components/navbar/navbar";
 
 const AdditionalPagesLayout = ({ children }: { children: React.ReactNode }) => {
   return (
