+++ app/(developer)/layout.tsx
@@ -5,7 +5,7 @@ import { redirect, useSearchParams } from "next/navigation";
 import { RedirectToSignIn } from "@clerk/nextjs";
 import { useConvexAuth } from "convex/react";
 
-import { Navbar } from "@/components/navbar";
+import { Navbar } from "@/components/navbar/navbar";
 import { Skeleton } from "@/components/ui/skeleton";
 import { useRoleCheck } from "@/hooks/use-role-check";
 import { AdminProvider } from "./admin/_components/adminprovider";
