+++ app/(main)/results/layout.tsx
@@ -1,4 +1,4 @@
-import { Navbar } from "@/components/navbar";
+import { Navbar } from "@/components/navbar/navbar";
 
 const ResultPageLayout = ({ children }: { children: React.ReactNode }) => {
   return (
