+++ app/not-found.tsx
@@ -5,7 +5,7 @@ import { useRouter } from "next/navigation";
 import { ArrowLeft, Home } from "lucide-react";
 
 import { Footer } from "@/components/footer";
-import { Navbar } from "@/components/navbar";
+import { Navbar } from "@/components/navbar/navbar";
 import { Button } from "@/components/ui/button";
 
 export default function NotFound() {
